+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Changed `onEachOperator` based MDC and trace propagation from `reactor.core.publisher.Hooks.onEachOperator` to `reactor.core.publisher.Hooks.addQueueWrapper` based implementation.
+- MDC and trace propagation always uses queue wrappers.  
+
+### Deprecated
+- Deprecated `com.hotels.molten.core.mdc.MoltenMDC.initialize(boolean)` in favour of `com.hotels.molten.core.mdc.MoltenMDC.initialize()`.
+- Deprecated `com.hotels.molten.trace.MoltenTrace.initialize(boolean)` in favour of `com.hotels.molten.trace.MoltenTrace.initialize()`.
 
 ## [1.2.1]
 ### Fixed